Integral acceleration - initial experiment
`LOOKAHEAD` is not supported with these changes.  It needs to be adjusted
to work with this method before this can be adopted into experimental.

The code works by calculating the velocity difference for each `TICK_TIME`.
This value is constant for constant acceleration.  Since it is a
fractional value, we use a fixed-point q8.24 representation to hold it.
This should work without overflow for most reasonable values of
`ACCELERATION` and `STEPS_PER_M`.

The adjusted velocity is used to calculate the new `dda->c` at each
`dda_clock` with a single call to `muldiv` during ramp-up and ramp-down.

Some variables are no longer needed, maybe, like dda->n.  But I didn't
try to reduce those yet since lookahead is still unresolved.